[jit] break out test_recursive_script
